MegaDesk (Independent) Submission
Created the MegaDesk project. Currently all forms exist, there's data validation for desk width and depth, a try-catch for FileNotFoundException in ViewQuotes.cs, and a successfully functioning Submit button for the Add Quote form. This submits all information to DeskQuote, which writes the data to a file. The DisplayQuote form automatically pops up to display this information, and the ViewQuotes form also reads this data successfully from a file. I've yet to code the SearchQuotes form, and probably need to optimize the program a fair bit more.